Add files via upload
The notebook for the final project. The images should be in separate subfolders of the data folder. At the same hierarchical level as the data folder, there should be a training_data and test_data folder. The code within the notebook will fill those out before the VGG and Inception models are fitted to the data. I will soon upload an archive containing the notebook file and the necessary folder structure and data to get started, for those who want to test this out for themselves.